figma-transformer: consolidate layout intent classification
